dev: refactor rest api
