+++ phpcs.baseline.xml
@@ -1972,45 +1972,6 @@
 <violation file="inc/prices.php" sniff="Squiz.PHP.CommentedOutCode.Found" signature="ff5432f8b4e66f4c3f51f7719d4d912089796f98"/>
 <violation file="inc/prices.php" sniff="Squiz.PHP.DisallowMultipleAssignments.FoundInControlStructure" signature="81d00833b2967ba5de60e4c0045bfe1aa8790112"/>
 <violation file="inc/prices.php" sniff="WordPress.NamingConventions.ValidVariableName.VariableNotSnakeCase" signature="9119234e03f190bcc7c6b484414ce2fd1535a832"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="074a7adc69719b11a61c85714b6d2eb09f622831"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="0f202bc684f4a05fa22bf2cce63eb8cdbfd30269"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="3075464b4d304076fb030e73519e8ccbf93fde9c"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="431b4e40ee98d28ed99669208377951f8674eb59"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="441e50d08f146eae1ca0dfd37eeb91230cf12a7f"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="4b4144da76e2543e48b00ece775c6c8ecf4e6c6e"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="5009007b60914dbbf76014753557bbe95b7bd129"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="54fcc9eaf7cb65b1d970720458f39fe3942d56df"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="5fa4daf48c3136b20901b417541c6b7c47281108"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="6117ad1b6ff8a2e8205fd2da79b26f6ded7f98f8"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="66f03030b96067a6a91dd21866b9a14904b114ec"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="68e118561d8c442ce2730e0b1d56b7c7bc70c7bf"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="6ccad6673b8156d45c4bde2cb793fb6ba51c2768"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="8500f8b1c75b404718e0f62cfc629b6bf0977128"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="8618a0a7b116250905bab2cbde1eaa7bbaceb9f8"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="8801d37730cebcccb2252fb4cbfe8513406e76f5"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="8b3e131c4f3d1003d7f8e7170b9862854718b231"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="926cfda014b26a96316c73af00dbabc8469afb1a"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="94828358d5b51b5f0c37d8f207e8813002687a1b"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="d6978a0eac53b14cd40f4f3e133627c5676eb8d4"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="d81426dc62a0af35aaa0e0dc5035893a3f09a545"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="dafb996edbd226ff5d77563a7a0f9e532032afaa"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="e29b633315b58676c0dfeb575105e2c73285d599"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="e2a4050445e99d5dfc1cf655a0f6af077fb0d757"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="e57cb2fa26ab6e70a3966bb56685ae68fbc2d99b"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="ef4f88b5f93d0ffe02767b95cfd1f263a8bf9c83"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="f4092048407c3ac32623262c42a0b5a8426ca38b"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="f7e62142afe1977721f8849d45ddc357869b153b"/>
-<violation file="inc/rest.class.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="fc470425721c89be2be17da981301dd63ea95d85"/>
-<violation file="inc/rest.class.php" sniff="Squiz.PHP.CommentedOutCode.Found" signature="6117ad1b6ff8a2e8205fd2da79b26f6ded7f98f8"/>
-<violation file="inc/rest.class.php" sniff="Squiz.PHP.CommentedOutCode.Found" signature="66f03030b96067a6a91dd21866b9a14904b114ec"/>
-<violation file="inc/rest.class.php" sniff="Squiz.PHP.CommentedOutCode.Found" signature="8801d37730cebcccb2252fb4cbfe8513406e76f5"/>
-<violation file="inc/rest.class.php" sniff="Squiz.PHP.CommentedOutCode.Found" signature="f7e62142afe1977721f8849d45ddc357869b153b"/>
-<violation file="inc/rest.class.php" sniff="WordPress.Security.ValidatedSanitizedInput.InputNotValidated" signature="d81426dc62a0af35aaa0e0dc5035893a3f09a545"/>
-<violation file="inc/rest.class.php" sniff="WordPress.Security.ValidatedSanitizedInput.InputNotValidatedNotSanitized" signature="2de17fc8a90b88fd1d9093db21e8e0a9ecd9c4bb"/>
-<violation file="inc/rest.class.php" sniff="WordPress.Security.ValidatedSanitizedInput.InputNotValidatedNotSanitized" signature="582d0af2d6a9aa7751bdeaea2869f3a9dcceee1a"/>
-<violation file="inc/rest.class.php" sniff="WordPress.WP.AlternativeFunctions.json_encode_json_encode" signature="c39c6f5294dd625e5a124af7e4d35ec3c791649e"/>
-<violation file="inc/rest.class.php" sniff="WordPress.WP.AlternativeFunctions.json_encode_json_encode" signature="d96124c6834c4c2b83d645fd919cfd7c58957d5d"/>
-<violation file="inc/rest.class.php" sniff="WordPress.WP.AlternativeFunctions.json_encode_json_encode" signature="fb61b8e2af691d90566b2095051e5cba7a785024"/>
 <violation file="inc/validation.php" sniff="Generic.CodeAnalysis.UnusedFunctionParameter.FoundAfterLastUsed" signature="3545362a5c17684d1e84abaf04b60e036b388b80"/>
 <violation file="inc/validation.php" sniff="Generic.CodeAnalysis.UnusedFunctionParameter.FoundAfterLastUsed" signature="7d1a5db4dc140323d0dfe566e5054aae4c0290ae"/>
 <violation file="inc/validation.php" sniff="Squiz.Commenting.InlineComment.InvalidEndChar" signature="08db6864b9205b4bc073d33bb9a3d2e4f555951d"/>
