+++ phpstan-baseline.neon
@@ -5590,97 +5590,31 @@ parameters:
 			message: '#^Access to an undefined property object\:\:\$productmeta_id\.$#'
 			identifier: property.notFound
 			count: 10
-			path: inc/rest.class.php
+			path: src/Rest/ProductFieldGroupService.php
 
 		-
 			message: '#^Access to an undefined property object\:\:\$the_meta\.$#'
 			identifier: property.notFound
 			count: 2
-			path: inc/rest.class.php
+			path: src/Rest/ProductFieldGroupService.php
 
 		-
 			message: '#^Call to an undefined method WC_Order_Item\:\:get_product_id\(\)\.$#'
 			identifier: method.notFound
-			count: 3
-			path: inc/rest.class.php
+			count: 1
+			path: src/Rest/OrderItemMetaPresenter.php
+
+		-
+			message: '#^Call to an undefined method WC_Order_Item\:\:get_product_id\(\)\.$#'
+			identifier: method.notFound
+			count: 2
+			path: src/Rest/OrderMetaController.php
 
 		-
 			message: '#^Cannot call method get_items\(\) on WC_Order\|WC_Order_Refund\|false\.$#'
 			identifier: method.nonObject
 			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:delete_meta_data\(\) has parameter \$delete_fields with no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:delete_meta_data\(\) return type has no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:filter_required_keys_only\(\) has parameter \$ppom_fields with no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:filter_required_keys_only\(\) return type has no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:get_order_item_meta\(\) return type has no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:save_new_meta_data\(\) has parameter \$ppom_fields with no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:save_new_meta_data\(\) return type has no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:update_meta_data\(\) has parameter \$ppom_fields with no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Method PPOM_Rest\:\:update_meta_data\(\) return type has no value type specified in iterable type array\.$#'
-			identifier: missingType.iterableValue
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Parameter \#1 \$ppom_meta of method PPOM_Rest\:\:delete_meta_data\(\) expects object, array\|object given\.$#'
-			identifier: argument.type
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Parameter \#1 \$ppom_meta of method PPOM_Rest\:\:update_meta_data\(\) expects object, array\|object given\.$#'
-			identifier: argument.type
-			count: 1
-			path: inc/rest.class.php
-
-		-
-			message: '#^Parameter \#1 \$str of function trim expects string, bool given\.$#'
-			identifier: argument.type
-			count: 1
-			path: inc/rest.class.php
+			path: src/Rest/OrderItemMetaPresenter.php
 
 		-
 			message: '#^Function ppom_fields_with_html\(\) return type has no value type specified in iterable type array\.$#'
